+++ src/main/java/in/projecteka/consentmanager/clients/model/KeycloakCreateUserRequest.java
@@ -1,10 +1,13 @@
-package in.projecteka.consentmanager.user.model;
+package in.projecteka.consentmanager.clients.model;
 
+import in.projecteka.consentmanager.user.model.UserCredential;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 
 import java.util.List;
 
+@Builder
 @Data
 @AllArgsConstructor
 public class KeycloakCreateUserRequest {
