+++ src/main/java/in/projecteka/consentmanager/user/UserConfiguration.java
@@ -3,8 +3,10 @@ package in.projecteka.consentmanager.user;
 import com.google.common.cache.CacheBuilder;
 import com.google.common.cache.CacheLoader;
 import com.google.common.cache.LoadingCache;
-import in.projecteka.consentmanager.clients.KeycloakClient;
+import in.projecteka.consentmanager.clients.IdentityServiceClient;
 import in.projecteka.consentmanager.clients.OtpServiceClient;
+import in.projecteka.consentmanager.clients.properties.IdentityServiceProperties;
+import in.projecteka.consentmanager.clients.properties.OtpServiceProperties;
 import in.projecteka.consentmanager.clients.properties.UserServiceProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -21,16 +23,14 @@ public class UserConfiguration {
                                    OtpServiceProperties otpServiceProperties,
                                    OtpServiceClient otpServiceClient,
                                    UserVerificationService userVerificationService,
-                                   KeycloakClient keycloakClient,
+                                   IdentityServiceClient identityServiceClient,
                                    TokenService tokenService) {
-        return new UserService(
-                userRepository,
+        return new UserService(userRepository,
                 otpServiceProperties,
                 otpServiceClient,
                 userVerificationService,
-                keycloakClient,
-                tokenService
-        );
+                identityServiceClient,
+                tokenService);
     }
 
     @Bean
@@ -45,9 +45,9 @@ public class UserConfiguration {
     }
 
     @Bean
-    public KeycloakClient keycloakClient(WebClient.Builder builder,
-                                         KeycloakProperties keycloakProperties) {
-        return new KeycloakClient(builder, keycloakProperties);
+    public IdentityServiceClient keycloakClient(WebClient.Builder builder,
+                                                IdentityServiceProperties identityServiceProperties) {
+        return new IdentityServiceClient(builder, identityServiceProperties);
     }
 
     @Bean
@@ -70,7 +70,7 @@ public class UserConfiguration {
     }
 
     @Bean
-    public TokenService tokenService(KeycloakProperties keycloakProperties, KeycloakClient keycloakClient) {
-        return new TokenService(keycloakProperties, keycloakClient);
+    public TokenService tokenService(IdentityServiceProperties identityServiceProperties, IdentityServiceClient identityServiceClient) {
+        return new TokenService(identityServiceProperties, identityServiceClient);
     }
 }
