+++ src/main/java/in/projecteka/consentmanager/user/model/SignUpRequest.java
@@ -1,8 +1,10 @@
 package in.projecteka.consentmanager.user.model;
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 
+@Builder
 @Data
 @AllArgsConstructor
 public class SignUpRequest {
