+++ src/main/java/in/projecteka/consentmanager/user/model/UserSignUpEnquiry.java
@@ -1,9 +1,11 @@
 package in.projecteka.consentmanager.user.model;
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+@Builder
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
